chore: refactored and improved approach
