+++ packages/el-cap-kit/test/el-cap-kit.test.js
@@ -1,7 +1,7 @@
 import test from 'node:test';
 import assert from 'node:assert/strict';
 import esmock from 'esmock';
-import { fetchRemainingData } from '../src/lib/el-cap-kit.js';
+import { fetchRemainingData } from '../src/lib/fetch-prices.js';
 
 test('should fetch data from Redstone API despite failure in fetching remaining data', async () => {
     const mockedData = { key1: 'value1', key2: 'value2', key3: 'value3' };
@@ -13,12 +13,13 @@ test('should fetch data from Redstone API despite failure in fetching remaining 
   
     // Set up a mock for `fetchRemainingData` and `fetchRedstonePrices`
     const mocked = await esmock('../src/lib/el-cap-kit.js', {
-      '../src/lib/fetch-prices.js': {
-        fetchData: fetchData,  // use our mocked fetchData;
-        fetchRedstonePrices: async () => mockedData,
-        fetchRemainingData: fetchRemainingData  // use real fetchRemainingData
-      },
-    });
+        '../src/lib/fetch-prices.js': {
+          fetchRedstonePrices: async () => mockedData,
+          fetchRemainingData: fetchRemainingData  // use real fetchRemainingData
+        },
+        'node-fetch': fetchData
+      });
+      
   
     let prices;
     try {
